requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
django-environ==0.4.5
django-extra-views==0.13.0
django-model-utils==4.1.1
django-widget-tweaks==1.4.8
psycopg2-binary==2.8.6
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
